+++ tests/template_ut/template_ut.c
@@ -353,7 +353,7 @@ BEGIN_TEST_SUITE(template_ut)
             umock_c_negative_tests_reset();
             umock_c_negative_tests_fail_call(i);
 
-            (void)sprintf(temp_str, "On failed call %zu", i);
+            (void)sprintf(temp_str, "On failed call %lu", (unsigned long)i);
 
             ///act
             result = target_create(SIZEOF_FOO_MEMORY);
@@ -403,7 +403,7 @@ BEGIN_TEST_SUITE(template_ut)
                 umock_c_negative_tests_reset();
                 umock_c_negative_tests_fail_call(i);
 
-                (void)sprintf(temp_str, "On failed call %zu", i);
+                (void)sprintf(temp_str, "On failed call %lu", (unsigned long)i);
 
                 ///act
                 result = target_create(SIZEOF_FOO_MEMORY);
@@ -452,7 +452,7 @@ BEGIN_TEST_SUITE(template_ut)
             umock_c_negative_tests_reset();
             umock_c_negative_tests_fail_call(i);
 
-            (void)sprintf(temp_str, "On failed call %zu", i);
+            (void)sprintf(temp_str, "On failed call %lu", (unsigned long)i);
 
             ///act
             result = target_foo();
